server/requirements.txt
asgiref==3.7.2
autopep8==2.0.1
backports.zoneinfo==0.2.1
distlib==0.3.6
Django==4.2.5
django-cors-headers==4.2.0
django-filter==23.3
django-ipware==5.0.2
django-modeltranslation==0.18.11
djangorestframework==3.14.0
drf-recaptcha==3.0.0
filelock==3.9.0
flake8==6.0.0
gunicorn==21.2.0
importlib-metadata==6.8.0
Markdown==3.4.4
mccabe==0.7.0
packaging==23.2
Pillow==10.0.1
platformdirs==3.0.0
pycodestyle==2.10.0
pyflakes==3.0.1
pytz==2023.3.post1
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.8.0
tzdata==2023.3
virtualenv==20.19.0
zipp==3.17.0
